feat(gymtrack): decommission legacy agent keys and unify on OAuth (task 1eb6e48c)
Replace the static-key GymTrack agent credential system (gymtrack_agent_api_keys)
with the MCP OAuth credential system. /api/agent/* endpoints now authenticate
bearer tokens against gymtrack_oauth_tokens + gymtrack_oauth_consents, and the
legacy key table, agentAuth.js helper, and planned_workouts.agent_key_id
column are all removed in favour of OAuth consent_id attribution.

- New apps/gymtrack/server/oauthAuth.js mirrors the legacy module shape and
  exports resolveOAuthIdentity(req, { requireScope }) which checks both the
  token row and the consent row for revoked_at and expiry on every request.
- Route handlers (history, planned-workouts, exercises/progression) now
  require per-method scopes (history:read, workouts:write, progression:read)
  and pass consent_id into createPlannedWorkout for attribution.
- Migration 20260815010000_decommission_legacy_agent_keys.sql revokes all
  existing legacy keys (audit-stamped), drops planned_workouts.agent_key_id,
  adds planned_workouts.consent_id (FK → gymtrack_oauth_consents, ON DELETE
  SET NULL), then drops the legacy table, indexes, and RLS policy.
- MCP server's plan_workout tool now threads consentId through to the
  GymTrack data layer instead of legacyAgentKeyId: null.
- Client-side plans.js + WorkoutsTab.test.jsx fixture switch from
  agent_key_id to consent_id in the row shape.
- apps/gymtrack/SPEC.md updated to describe OAuth-only agent credentials
  and rename the section to 'OAuth-protected REST endpoints'.
- Tests: new oauthAuth.test.js covers token/consent lookup, revocation,
  expiry, scope enforcement, and adminClient env validation; existing
  handler tests retarget the mock to oauthAuth.js with the new
  { user_id, consent_id, client_id, scope } identity shape; plannedWorkouts
  unit tests verify consent_id is recorded on the parent row.
