+++ apps/gymtrack/api/agent/exercises/[exerciseName]/progression.js
@@ -34,9 +34,9 @@ import {
   adminClient,
   badRequest,
   rejectIfWrongMethod,
-  resolveAgentIdentity,
+  resolveOAuthIdentity,
   unauthorized
-} from '../../../../server/agentAuth.js';
+} from '../../../../server/oauthAuth.js';
 import {
   escapeIlikePattern,
   exerciseNameErrorMessage,
@@ -67,7 +67,7 @@ export default async function handler(req, res) {
 
   let identity;
   try {
-    identity = await resolveAgentIdentity(req);
+    identity = await resolveOAuthIdentity(req, { requireScope: 'progression:read' });
   } catch (err) {
     return res.status(500).json({ error: 'server_error', message: err?.message ?? 'Auth lookup failed.' });
   }
